+++ tests/components/plugwise/test_button.py
@@ -2,15 +2,14 @@
 
 from unittest.mock import MagicMock
 
-from homeassistant.components.button import ButtonDeviceClass
-from homeassistant.const import (
-    ATTR_DEVICE_CLASS,
-    ATTR_ENTITY_ID,
-    STATE_UNKNOWN,
-    Platform,
+from homeassistant.components.button import (
+    DOMAIN as BUTTON_DOMAIN,
+    SERVICE_PRESS,
+    ButtonDeviceClass,
 )
+from homeassistant.const import ATTR_DEVICE_CLASS, ATTR_ENTITY_ID, STATE_UNKNOWN
 from homeassistant.core import HomeAssistant
-from homeassistant.helpers import entity_registry as er
+import homeassistant.helpers.entity_registry as er
 
 from tests.common import MockConfigEntry
 
@@ -30,8 +29,8 @@ async def test_adam_reboot_button(
     assert entry.unique_id == "fe799307f1624099878210aa0b9f1475-reboot"
 
     await hass.services.async_call(
-        Platform.BUTTON,
-        "press",
+        BUTTON_DOMAIN,
+        SERVICE_PRESS,
         {ATTR_ENTITY_ID: "button.adam_reboot"},
         blocking=True,
     )