+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Versions from 0.40 and up
 
+## Ongoing
+
+- Test-code improvements.
+
 ## v0.53.2
 
 - Climate-test: implement freezer.tick constructs.